requirements.txt
#--index-url https://download.pytorch.org/whl/cu118
#torch==2.0.0
stable-baselines3[extra]==1.7.0
sb3-contrib==1.7.0
fire==0.5.0
nip-config==0.8.3
